% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctr_keyfig.R
\name{plot_ctr_keyfig}
\alias{plot_ctr_keyfig}
\title{Summary Key Figures
 Key figures can be highlighted with humanitarian icons
  https://fontawesome.com/icons/categories/humanitarian}
\usage{
plot_ctr_keyfig(
  year = 2024,
  country_asylum_iso3c,
  population_type_font_size = 4,
  population_size_font_size = 5,
  icon_size = 5
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{country_asylum_iso3c}{Character value with the ISO-3 character code of the Country of Asylum}

\item{population_type_font_size}{Numeric value for population type font size, default to 4}

\item{population_size_font_size}{Numeric value for population size number font size, default to 5}

\item{icon_size}{Numeric value for icon size, default to 5}
}
\value{
ggplot2
}
\description{
Summary Key Figures
 Key figures can be highlighted with humanitarian icons
  https://fontawesome.com/icons/categories/humanitarian
}
\examples{
plot_ctr_keyfig(
  year = 2024,
  country_asylum_iso3c = "COL"
)
plot_ctr_keyfig(
  year = 2024,
  country_asylum_iso3c = "COL"
)
}
