% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctr_population_type_abs.R
\name{plot_ctr_population_type_abs}
\alias{plot_ctr_population_type_abs}
\title{Main country of origin - Absolute value}
\usage{
plot_ctr_population_type_abs(
  year = 2024,
  country_asylum_iso3c,
  top_n_countries = 9,
  pop_type = "REF",
  show_diff_label = TRUE,
  label_font_size = 4,
  category_font_size = 10
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{country_asylum_iso3c}{Character value with the ISO-3 character code of the Country of Asylum}

\item{top_n_countries}{Numeric value of number of main countries that the graph should display}

\item{pop_type}{Character value. Possible population type (e.g.: REF, IDP, ASY,   OIP, OOC, STA)}

\item{show_diff_label}{logical to indicate whether or not adding the the label displaying difference in percentage compared to the previous year}

\item{label_font_size}{Numeric value for label font size, default to 4}

\item{category_font_size}{Numeric value for axis text font size, default to 10}
}
\value{
a ggplot2 object
}
\description{
Main country of origin - Absolute value
}
\examples{
plot_ctr_population_type_abs(
  year = 2024,
  country_asylum_iso3c = "USA",
  top_n_countries = 4,
  pop_type = "REF",
  show_diff_label = FALSE,
  label_font_size = 4,
  category_font_size = 10
)

## Same with 9 top countries and Asylum-seekers included
plot_ctr_population_type_abs(
  year = 2024,
  country_asylum_iso3c = "USA",
  top_n_countries = 9,
  pop_type = "ASY",
  show_diff_label = TRUE,
  label_font_size = 4,
  category_font_size = 10
)
plot_ctr_population_type_abs(
  year = 2024,
  country_asylum_iso3c = "USA",
  top_n_countries = 4,
  pop_type = "REF",
  show_diff_label = FALSE
)

## Same with 9 top countries and Asylum-seekers included
plot_ctr_population_type_abs(
  year = 2024,
  country_asylum_iso3c = "USA",
  top_n_countries = 9,
  pop_type = "ASY",
  show_diff_label = TRUE
)
}
