% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctr_solution_recognition.R
\name{plot_ctr_solution_recognition}
\alias{plot_ctr_solution_recognition}
\title{Title}
\usage{
plot_ctr_solution_recognition(
  year = 2024,
  lag = 10,
  country_asylum_iso3c = country_asylum_iso3c,
  category_font_size = 10
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{lag}{Number of year to used as comparison base}

\item{country_asylum_iso3c}{Character value with the ISO-3 character code of the Country of Asylum}

\item{category_font_size}{Numeric value for axis text font size, default to 10}
}
\value{
a ggplot2 object
}
\description{
Description
}
\examples{
plot_ctr_solution_recognition(
  year = 2024,
  country_asylum_iso3c = "UGA",
  lag = 10,
  category_font_size = 10
)
plot_ctr_solution_recognition(
  year = 2024,
  country_asylum_iso3c = "UGA",
  lag = 10
)
}
