% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctr_treemap.R
\name{plot_ctr_treemap}
\alias{plot_ctr_treemap}
\title{Tree map of Population Groups within a country}
\usage{
plot_ctr_treemap(
  year = 2024,
  country_asylum_iso3c = country_asylum_iso3c,
  label_font_size = 12
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{country_asylum_iso3c}{Character value with the ISO-3 character code of the Country of Asylum}

\item{label_font_size}{Numeric value for label font size, default to 12}
}
\value{
a ggplot2 object
}
\description{
Tree map of Population Groups within a country
}
\examples{
#
plot_ctr_treemap(
  year = 2024,
  country_asylum_iso3c = "USA",
  label_font_size = 12
)
#
plot_ctr_treemap(
  year = 2024,
  country_asylum_iso3c = "USA"
)
}
