% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_decrease.R
\name{plot_reg_decrease}
\alias{plot_reg_decrease}
\title{Plot Regional Decrease}
\usage{
plot_reg_decrease(
  year = 2024,
  lag = 5,
  topn = 5,
  region = "The Americas",
  pop_type = c("REF", "ASY", "OIP"),
  category_font_size = 10
)
}
\description{
Plot Regional Decrease
}
\examples{
plot_reg_decrease(
    year = 2024,
    region = "The Americas",
    category_font_size = 10
)
}
