% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_increase.R
\name{plot_reg_increase}
\alias{plot_reg_increase}
\title{Plot Regional Increase}
\usage{
plot_reg_increase(
  year = 2024,
  lag = 5,
  topn = 5,
  region = "The Americas",
  pop_type = c("REF", "ASY", "OIP"),
  category_font_size = 10
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{lag}{Number of year to used as comparison base}

\item{topn}{Numeric value for number of top countries to show}

\item{region}{Character value including the UNHCR region}

\item{pop_type}{Vector of character values. Possible population type}

\item{category_font_size}{Numeric value for axis text font size}
}
\description{
Plot Regional Increase
}
\examples{
plot_reg_increase(
    year = 2024,
    region = "The Americas",
    category_font_size = 10
)
}
