% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_map.R
\name{plot_reg_map}
\alias{plot_reg_map}
\title{Plot Regional Map}
\usage{
plot_reg_map(
  year = 2024,
  region = "The Americas",
  topn = 5,
  pop_type = c("REF", "ASY", "OIP"),
  projection = "Mercator",
  maxSymbolsize = 0.25
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{region}{Character value including the UNHCR region}

\item{topn}{Numeric value for number of top items}

\item{pop_type}{Vector of character values. Possible population type}

\item{projection}{Character value for map projection}

\item{maxSymbolsize}{Numeric value for maximum symbol size}
}
\description{
Plot Regional Map
}
\examples{
plot_reg_map(
    year = 2024,
    region = "The Americas"
)
}
