% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_population_type_abs.R
\name{plot_reg_population_type_abs}
\alias{plot_reg_population_type_abs}
\title{Plot Regional Population Type Absolute}
\usage{
plot_reg_population_type_abs(
  year = 2024,
  region = "The Americas",
  top_n_countries = 5,
  pop_type = "REF",
  show_diff_label = TRUE,
  label_font_size = 6,
  category_font_size = 10
)
}
\description{
Plot Regional Population Type Absolute
}
\examples{
plot_reg_population_type_abs(
    year = 2024,
    region = "The Americas",
    label_font_size = 6,
    category_font_size = 10
)
}
