% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_population_type_per_year.R
\name{plot_reg_population_type_per_year}
\alias{plot_reg_population_type_per_year}
\title{Plot Regional Population Type per Year}
\usage{
plot_reg_population_type_per_year(
  year = 2024,
  region = "The Americas",
  lag = 5,
  pop_type = c("REF", "ASY", "IDP", "OIP", "STA", "OOC"),
  label_font_size = 4,
  category_font_size = 10,
  legend_font_size = 10
)
}
\description{
Plot Regional Population Type per Year
}
\examples{
plot_reg_population_type_per_year(
    year = 2024,
    region = "The Americas",
    label_font_size = 4,
    category_font_size = 10,
    legend_font_size = 10
)
}
