% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_prop_origin.R
\name{plot_reg_prop_origin}
\alias{plot_reg_prop_origin}
\title{Plot Regional Proportion Origin}
\usage{
plot_reg_prop_origin(
  year = 2024,
  region = "The Americas",
  label_font_size = 6,
  category_font_size = 10
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{region}{Character value including the UNHCR region}

\item{label_font_size}{Numeric value for label font size}

\item{category_font_size}{Numeric value for axis text font size}
}
\description{
Plot Regional Proportion Origin
}
\examples{
plot_reg_prop_origin(
    year = 2024,
    region = "The Americas",
    label_font_size = 6,
    category_font_size = 10
)
}
