% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_rsd.R
\name{plot_reg_rsd}
\alias{plot_reg_rsd}
\title{Plot Regional RSD}
\usage{
plot_reg_rsd(
  year = 2024,
  region,
  top_n_countries = 10,
  measure = "Recognized",
  category_font_size = 10
)
}
\description{
Plot Regional RSD
}
\examples{
plot_reg_rsd(
    year = 2024,
    region = "The Americas",
    measure = "Recognized",
    category_font_size = 10
)
}
