% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_share.R
\name{plot_reg_share}
\alias{plot_reg_share}
\title{Plot Regional Share}
\usage{
plot_reg_share(
  year = 2024,
  region = "The Americas",
  pop_type = "REF",
  label_font_size = 15
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{region}{Character value including the UNHCR region}

\item{pop_type}{Vector of character values. Possible population type}

\item{label_font_size}{Numeric value for label font size}
}
\description{
Plot Regional Share
}
\examples{
plot_reg_share(
    year = 2024,
    region = "The Americas",
    label_font_size = 15
)
}
