% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_treemap.R
\name{plot_reg_treemap}
\alias{plot_reg_treemap}
\title{Plot Regional Treemap}
\usage{
plot_reg_treemap(
  year = 2024,
  region = "The Americas",
  pop_type = c("REF", "ASY", "IDP", "OIP", "OOC", "STA"),
  label_font_size = 15
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{region}{Character value including the UNHCR region}

\item{pop_type}{Vector of character values. Possible population type (e.g.: REF, IDP, ASY, OIP, OIP, OOC, STA)}

\item{label_font_size}{Numeric value for label font size}
}
\description{
Plot Regional Treemap
}
\examples{
plot_reg_treemap(
    year = 2024,
    region = "The Americas",
    label_font_size = 15
)
}
