% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_map.R
\name{plot_reg_map}
\alias{plot_reg_map}
\title{Plot Regional Map}
\usage{
plot_reg_map(
  year = 2024,
  region = "The Americas",
  topn = 5,
  pop_type = c("REF", "ASY", "OIP"),
  projection = "Mercator",
  maxSymbolsize = 0.25
)
}
\description{
Plot Regional Map
}
\examples{
plot_reg_map(
    year = 2024,
    region = "The Americas"
)
}
