% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_origin_dest.R
\name{plot_reg_origin_dest}
\alias{plot_reg_origin_dest}
\title{Plot Regional Origin Destination}
\usage{
plot_reg_origin_dest(year = 2024, region = "The Americas")
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{region}{Character value including the UNHCR region}
}
\description{
Plot Regional Origin Destination
}
\examples{
plot_reg_origin_dest(
    year = 2024,
    region = "The Americas"
)
}
