% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_population_type_abs.R
\name{plot_reg_population_type_abs}
\alias{plot_reg_population_type_abs}
\title{Plot Regional Population Type Absolute}
\usage{
plot_reg_population_type_abs(
  year = 2024,
  region = "The Americas",
  top_n_countries = 5,
  pop_type = "REF",
  show_diff_label = TRUE,
  label_font_size = 6,
  category_font_size = 10
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{region}{Character value including the UNHCR region}

\item{top_n_countries}{Numeric value for number of top countries to show}

\item{pop_type}{Vector of character values. Possible population type}

\item{show_diff_label}{Boolean to show difference label}

\item{label_font_size}{Numeric value for label font size}

\item{category_font_size}{Numeric value for axis text font size}
}
\description{
Plot Regional Population Type Absolute
}
\examples{
plot_reg_population_type_abs(
    year = 2024,
    region = "The Americas",
    label_font_size = 6,
    category_font_size = 10
)
}
