% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_rsd.R
\name{plot_reg_rsd}
\alias{plot_reg_rsd}
\title{Plot Regional RSD}
\usage{
plot_reg_rsd(
  year = 2024,
  region,
  top_n_countries = 10,
  measure = "Recognized",
  category_font_size = 10
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{region}{Character value including the UNHCR region}

\item{top_n_countries}{Numeric value for number of top countries to show}

\item{measure}{Character value for the measure to display}

\item{category_font_size}{Numeric value for axis text font size}
}
\description{
Plot Regional RSD
}
\examples{
plot_reg_rsd(
    year = 2024,
    region = "The Americas",
    measure = "Recognized",
    category_font_size = 10
)
}
