% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_solution.R
\name{plot_reg_solution}
\alias{plot_reg_solution}
\title{Plot Regional Solution}
\usage{
plot_reg_solution(
  year = 2024,
  region = "The Americas",
  lag = 10,
  label_font_size = 4,
  category_font_size = 10
)
}
\arguments{
\item{year}{Numeric value of the year (for instance 2020)}

\item{region}{Character value including the UNHCR region}

\item{lag}{Number of year to used as comparison base}

\item{label_font_size}{Numeric value for label font size}

\item{category_font_size}{Numeric value for axis text font size}
}
\description{
Plot Regional Solution
}
\examples{
plot_reg_solution(
    year = 2024,
    region = "The Americas",
    label_font_size = 4,
    category_font_size = 10
)
}
