% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_treemap.R
\name{plot_reg_treemap}
\alias{plot_reg_treemap}
\title{Plot Regional Treemap}
\usage{
plot_reg_treemap(
  year = 2024,
  region = "The Americas",
  pop_type = c("REF", "ASY", "IDP", "OIP", "OOC", "STA"),
  label_font_size = 15
)
}
\description{
Plot Regional Treemap
}
\examples{
plot_reg_treemap(
    year = 2024,
    region = "The Americas",
    label_font_size = 15
)
}
